requirements.txt
ariadne
graphql-core
python-dateutil
pyodbc